parallelcut modification

+++ equipe45/src/main/java/Equipe45/domain/CNC.java
@@ -96,7 +96,7 @@ public class CNC {
         }
     }
     
-    public void ModifyDistance(float distance){
+    public void ModifyDistance(int distance){
         if(distance < 0) 
             return;
